+++ allensdk/test/model/peri_model/modfiles/CaDynamics.mod
@@ -1,0 +1,40 @@
+: Dynamics that track inside calcium concentration
+: modified from Destexhe et al. 1994
+
+NEURON	{
+	SUFFIX CaDynamics
+	USEION ca READ ica WRITE cai
+	RANGE decay, gamma, minCai, depth
+}
+
+UNITS	{
+	(mV) = (millivolt)
+	(mA) = (milliamp)
+	FARADAY = (faraday) (coulombs)
+	(molar) = (1/liter)
+	(mM) = (millimolar)
+	(um)	= (micron)
+}
+
+PARAMETER	{
+	gamma = 0.05 : percent of free calcium (not buffered)
+	decay = 80 (ms) : rate of removal of calcium
+	depth = 0.1 (um) : depth of shell
+	minCai = 1e-4 (mM)
+}
+
+ASSIGNED	{ica (mA/cm2)}
+
+INITIAL {
+	cai = minCai
+}
+
+STATE	{
+	cai (mM)
+}
+
+BREAKPOINT	{ SOLVE states METHOD cnexp }
+
+DERIVATIVE states	{
+	cai' = -(10000)*(ica*gamma/(2*FARADAY*depth)) - (cai - minCai)/decay
+}
