+++ allensdk/test/model/peri_model/modfiles/K_T.mod
@@ -1,0 +1,68 @@
+: Comment: The transient component of the K current
+: Reference:		Voltage-gated K+ channels in layer 5 neocortical pyramidal neurones from young rats:subtypes and gradients,Korngreen and Sakmann, J. Physiology, 2000
+
+NEURON	{
+	SUFFIX K_T
+	USEION k READ ek WRITE ik
+	RANGE gbar, g, ik
+}
+
+UNITS	{
+	(S) = (siemens)
+	(mV) = (millivolt)
+	(mA) = (milliamp)
+}
+
+PARAMETER	{
+	gbar = 0.00001 (S/cm2)
+	vshift = 0 (mV)
+	mTauF = 1.0
+	hTauF = 1.0
+}
+
+ASSIGNED	{
+	v	(mV)
+	ek	(mV)
+	ik	(mA/cm2)
+	g	(S/cm2)
+	celsius (degC)
+	mInf
+	mTau
+	hInf
+	hTau
+}
+
+STATE	{
+	m
+	h
+}
+
+BREAKPOINT	{
+	SOLVE states METHOD cnexp
+	g = gbar*m*m*m*m*h
+	ik = g*(v-ek)
+}
+
+DERIVATIVE states	{
+	rates()
+	m' = (mInf-m)/mTau
+	h' = (hInf-h)/hTau
+}
+
+INITIAL{
+	rates()
+	m = mInf
+	h = hInf
+}
+
+PROCEDURE rates(){
+  LOCAL qt
+  qt = 2.3^((celsius-21)/10)
+
+	UNITSOFF
+		mInf =  1/(1 + exp(-(v - (-47 + vshift)) / 29))
+		mTau =  (0.34 + mTauF * 0.92*exp(-((v+71-vshift)/59)^2))/qt
+		hInf =  1/(1 + exp(-(v+66-vshift)/-10))
+		hTau =  (8 + hTauF * 49*exp(-((v+73-vshift)/23)^2))/qt
+	UNITSON
+}
