+++ allensdk/test/model/peri_model/modfiles/SK.mod
@@ -1,0 +1,56 @@
+: SK-type calcium-activated potassium current
+: Reference : Kohler et al. 1996
+
+NEURON {
+       SUFFIX SK
+       USEION k READ ek WRITE ik
+       USEION ca READ cai
+       RANGE gbar, g, ik
+}
+
+UNITS {
+      (mV) = (millivolt)
+      (mA) = (milliamp)
+      (mM) = (milli/liter)
+}
+
+PARAMETER {
+          v            (mV)
+          gbar = .000001 (mho/cm2)
+          zTau = 1              (ms)
+          ek           (mV)
+          cai          (mM)
+}
+
+ASSIGNED {
+         zInf
+         ik            (mA/cm2)
+         g	       (S/cm2)
+}
+
+STATE {
+      z   FROM 0 TO 1
+}
+
+BREAKPOINT {
+           SOLVE states METHOD cnexp
+           g  = gbar * z
+           ik   =  g * (v - ek)
+}
+
+DERIVATIVE states {
+        rates(cai)
+        z' = (zInf - z) / zTau
+}
+
+PROCEDURE rates(ca(mM)) {
+          if(ca < 1e-7){
+	              ca = ca + 1e-07
+          }
+          zInf = 1/(1 + (0.00043 / ca)^4.8)
+}
+
+INITIAL {
+        rates(cai)
+        z = zInf
+}
